宣传用品管理-印刷品,宣传画面设计 物料制作 导入 [ci skip]

+++ src/main/java/com/hbhb/cw/publicity/web/vo/PrintInitVO.java
@@ -16,7 +16,7 @@ import java.io.Serializable;
 public class PrintInitVO implements Serializable {
     private static final long serialVersionUID = -2614659372873662253L;
     @Schema(description = "宣传印刷品id")
-    Integer printId;
+    Long printId;
     @Schema(description = "流程类型id")
     Long flowTypeId;
     @Schema(description = "用户id")
